requirements.txt
pandas==2.0.3
ipykernel==6.29.5
scikit-learn==1.3.2
scikit-surprise==1.1.4
matplotlib==3.7.5
xgboost==2.1.3
optuna==4.1.0
shap==0.44.1